feat: AuditorAware, PageInfo

+++ src/main/java/com/example/proto/domain/organization/department/DepartmentRepository.java
@@ -2,5 +2,5 @@ package com.example.proto.domain.organization.department;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartmentRepository extends JpaRepository<Department, Long> {
+public interface DepartmentRepository extends JpaRepository<Department, Integer> {
 }
